refactor: /constants/enums enum을 관리하는 폴더 추가
