+++ backend/src/main/java/com/project/Tamago/constants/enums/KeyType.java
@@ -1,4 +1,4 @@
-package com.project.Tamago.constants;
+package com.project.Tamago.constants.enums;
 
 import lombok.Getter;
