+++ backend/src/main/java/com/project/Tamago/constants/enums/Mode.java
@@ -1,0 +1,5 @@
+package com.project.Tamago.constants.enums;
+
+public enum Mode {
+	PRACTICE, ACTUAL
+}
