+++ backend/src/main/java/com/project/Tamago/domain/User.java
@@ -13,7 +13,7 @@ import org.hibernate.annotations.DynamicInsert;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import com.project.Tamago.dto.requestDto.ModifyProfileReqDto;
-import com.project.Tamago.constants.Role;
+import com.project.Tamago.constants.enums.Role;
 import com.project.Tamago.constants.converter.RoleConverter;
 
 import lombok.AccessLevel;
