--- backend/src/main/java/com/project/Tamago/domain/enums/Mode.java
@@ -1,5 +1,0 @@
-package com.project.Tamago.domain.enums;
-
-public enum Mode {
-	PRACTICE, ACTUAL
-}
